pricing/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
stripe==11.3.0
firebase-admin==6.6.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
google-cloud-scheduler==2.14.1
pydantic==2.10.3
email-validator==2.2.0
python-dotenv==1.0.1
httpx==0.28.1
